split net module into multiple files
